Commit APIs Route fix bug

+++ Backend/database/migrations/2024_12_24_014233_add_upload_count_to_users_table.php
@@ -9,14 +9,26 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('upload_count')->default(0);
+            if (!Schema::hasColumn('users', 'upload_count')) {
+                $table->integer('upload_count')->default(0);
+            }
+
+            if (!Schema::hasColumn('users', 'image_profile')) {
+                $table->string('image_profile')->nullable()->after('email');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('upload_count');
+            if (Schema::hasColumn('users', 'image_profile')) {
+                $table->dropColumn('image_profile');
+            }
+
+            if (Schema::hasColumn('users', 'upload_count')) {
+                $table->dropColumn('upload_count');
+            }
         });
     }
 };
